requirements.txt
langchain 
python-dotenv 
streamlit 
unstructured 
tiktoken 
faiss-cpu 
libmagic 
python-magic 
OpenAI
langchain_community